fix: [sc-15] move send mail function from utils to helpers folder
