+++ src/helpers/mail.helper.js
@@ -1,10 +1,28 @@
-const sendMail = require('../utils/mail.util.js');
+const nodemailer = require('nodemailer');
+
+const transporter = nodemailer.createTransport({
+  service: 'gmail',
+  host: process.env.MAIL_HOST,
+  port: process.env.MAIL_PORT,
+  auth: {
+    user: process.env.EMAIL_USER,
+    pass: process.env.EMAIL_PASSWORD,
+  },
+});
 
 const sendOtpEmail = async (email, otp) => {
+  const mailOptions = {
+    from: process.env.EMAIL_USER,
+    to: email,
+    subject: 'Verification Code',
+    text: `Your OTP code is ${otp}`,
+  };
   console.log(`email in mail helper: ${email}`);
-  const subject = 'Your OTP Code';
-  const text = `Your OTP code is ${otp}`;
-  return sendMail(email, subject, text);
+  try {
+    await transporter.sendMail(mailOptions);
+  } catch (error) {
+    console.error('Email send error:', error);
+  }
 };
 
 module.exports = sendOtpEmail;
